perf: extract every SessionBar item into its own memoized component
Continues #536's pattern (DriverBadgeItem/SofItem/ClassDriversItem) for
the remaining items. The parent SessionBar previously held ~14 hook
subscriptions at the top level — useBrakeBias, useDriverIncidents,
useSessionLapCount, useTrackTemperature, useCurrentTime, etc. — and
re-rendered at telemetry frequency regardless of which items were
enabled. With 5 SessionBar instances stacked across Standings/Relative/
InformationBar header+footer, that's a lot of wasted work.

Each item now lives in its own memo'd component file and subscribes
only to the data it needs. The parent SessionBar becomes a thin layout
shell that subscribes to useGeneralSettings, maps displayOrder to the
mounted item components, and handles standalone first/last alignment.
Disabled items pay nothing — no telemetry subscription, no render.

New components: SessionNameItem, SessionTimeItem, SessionLapsItem,
IncidentCountItem, BrakeBiasItem, LocalTimeItem, SessionClockTimeItem,
TrackWetnessItem, PrecipitationItem, AirTemperatureItem,
TrackTemperatureItem, TrackNameItem, WindItem, HumidityItem.

SessionBar.tsx: 470 → 165 lines, behavior preserved. Lint and 760
tests pass.

Caveat: SessionTimeItem and SessionLapsItem both call useSessionLapCount
independently. When both are enabled (common in headers) the hook runs
twice instead of once. A follow-up could lift its state-tracking effect
into a Zustand store so it runs once globally.
