+++ src/frontend/components/Standings/components/SessionBar/IncidentCountItem.tsx
@@ -1,0 +1,13 @@
+import { memo } from 'react';
+import { useDriverIncidents } from '../../hooks';
+
+export const IncidentCountItem = memo(() => {
+  const { incidentLimit, incidents } = useDriverIncidents();
+  return (
+    <div className="flex justify-end">
+      {incidents}
+      {incidentLimit ? ' / ' + incidentLimit : ''} x
+    </div>
+  );
+});
+IncidentCountItem.displayName = 'IncidentCountItem';
