+++ src/frontend/components/Standings/components/SessionBar/SessionNameItem.tsx
@@ -1,0 +1,8 @@
+import { memo } from 'react';
+import { useCurrentSessionType } from '@irdashies/context';
+
+export const SessionNameItem = memo(() => {
+  const session = useCurrentSessionType();
+  return <div className="flex">{session}</div>;
+});
+SessionNameItem.displayName = 'SessionNameItem';
