+++ src/frontend/components/Standings/components/SessionBar/TrackNameItem.tsx
@@ -1,0 +1,8 @@
+import { memo } from 'react';
+import { useTrackDisplayName } from '@irdashies/context';
+
+export const TrackNameItem = memo(() => {
+  const trackDisplayName = useTrackDisplayName();
+  return <div className="flex">{trackDisplayName}</div>;
+});
+TrackNameItem.displayName = 'TrackNameItem';
